Correct casting of TID
git-svn-id: file:///home/jj/hercules.svn/trunk@7156 956126f8-22a0-4046-8f4a-272fa8102e63

+++ config.c
@@ -414,7 +414,7 @@ int configure_capping(U32 value)
 
         if ( cnt == 0 )
         {
-            WRMSG( HHC00105, "E", sysblk.captid, "Capping manager" );
+            WRMSG( HHC00105, "E", (u_long)sysblk.captid, "Capping manager" );
             return HERRTHREADACT;
         }
